Refactoring common utilities for datapipes
